requirements.txt
Django
django-classy-tags
django-coverage
